# docs/swagger.yaml
definitions:
  api.Response:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
    type: object
  models.Domain:
    properties:
      created_at:
        type: string
      name:
        type: string
      updated_at:
        type: string
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
      uuid:
        type: string
    type: object
  models.Post:
    properties:
      content:
        type: string
      created_at:
        type: string
      domain:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      userID:
        type: integer
      uuid:
        type: string
    type: object
  models.Role:
    properties:
      created_at:
        type: string
      domainID:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      uuid:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      domains:
        items:
          $ref: '#/definitions/models.Domain'
        type: array
      email:
        type: string
      name:
        type: string
      posts:
        items:
          $ref: '#/definitions/models.Post'
        type: array
      updated_at:
        type: string
      uuid:
        type: string
    type: object
...
  requests.RegisterRequest:
    properties:
      email:
        example: username@gmail.com
        type: string
      name:
        example: John Doe
        type: string
      password:
        example: P@ssw0r6
        type: string
    required:
    - email
    - name
    - password
    type: object
  requests.RoleRequest:
    properties:
      name:
        example: Operator
        type: string
    required:
    - name
    type: object
...
  requests.UpdateRequest:
    properties:
      name:
        example: John Doe
        type: string
    required:
    - name
    type: object
...
info:
  contact: {}
paths:
...
  /api/role:
    get:
      consumes:
      - application/json
      description: Returns a list of roles for the specified domain.
      operationId: role-list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Role'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: List roles
      tags:
      - Role Management
    post:
      consumes:
      - application/json
      description: Creates a new role in the specified domain.
      operationId: role-create
      parameters:
      - description: Role creation data
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.RoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Create role
      tags:
      - Role Management
  /api/role/{uuid}:
    delete:
      consumes:
      - application/json
      description: Removes a role by UUID from the specified domain.
      operationId: role-delete
      parameters:
      - description: Role UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete role
      tags:
      - Role Management
    get:
      consumes:
      - application/json
      description: Returns the details of a role by UUID within the specified domain.
      operationId: role-read
      parameters:
      - description: Role UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Role'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Get role
      tags:
      - Role Management
    put:
      consumes:
      - application/json
      description: Modifies the details of a role by UUID within the specified domain.
      operationId: role-update
      parameters:
      - description: Role UUID
        in: path
        name: uuid
        required: true
        type: string
      - description: Role update data
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.RoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Role'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Update role
      tags:
      - Role Management
  /api/user:
    get:
      consumes:
      - application/json
      description: Returns a list of users for the specified domain.
      operationId: user-list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: List users
      tags:
      - User Management
    post:
      consumes:
      - application/json
      description: Creates a new user in the specified domain.
      operationId: user-create
      parameters:
      - description: User registration data
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Create user
      tags:
      - User Management
  /api/user/{uuid}:
    delete:
      consumes:
      - application/json
      description: Removes a user by UUID from the specified domain.
      operationId: user-delete
      parameters:
      - description: User UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete user
      tags:
      - User Management
    get:
      consumes:
      - application/json
      description: Returns the details of a user by UUID within the specified domain.
      operationId: user-read
      parameters:
      - description: User UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Get user
      tags:
      - User Management
    put:
      consumes:
      - application/json
      description: Modifies the details of a user by UUID within the specified domain.
      operationId: user-update
      parameters:
      - description: User UUID
        in: path
        name: uuid
        required: true
        type: string
      - description: User update data
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Update user
      tags:
      - User Management
...
securityDefinitions:
  ApiKeyAuth:
    description: 'Provide the accessToken as a Bearer token: ''Bearer {accessToken}'''
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"